+++ sdk/keyvault/Azure.Security.KeyVault.Keys/samples/Sample1_HelloWorld.cs
@@ -39,12 +39,12 @@ namespace Azure.Security.KeyVault.Keys.Samples
 
             // Let's Get the Cloud RSA Key from the Key Vault.
             KeyVaultKey cloudRsaKey = client.GetKey(rsaKeyName);
-            Debug.WriteLine($"Key is returned with name {cloudRsaKey.Name} and type {cloudRsaKey.Key.KeyType}");
+            Debug.WriteLine($"Key is returned with name {cloudRsaKey.Name} and type {cloudRsaKey.KeyType}");
 
             // After one year, the Cloud RSA Key is still required, we need to update the expiry time of the key.
             // The update method can be used to update the expiry attribute of the key.
             cloudRsaKey.Properties.ExpiresOn.Value.AddYears(1);
-            KeyVaultKey updatedKey = client.UpdateKeyProperties(cloudRsaKey.Properties, cloudRsaKey.Key.KeyOps);
+            KeyVaultKey updatedKey = client.UpdateKeyProperties(cloudRsaKey.Properties, cloudRsaKey.KeyOperations);
             Debug.WriteLine($"Key's updated expiry time is {updatedKey.Properties.ExpiresOn}");
 
             // We need the Cloud RSA key with bigger key size, so you want to update the key in Key Vault to ensure